elements/card, dl, styles, components/tx/createTxForm: refactor styles for reuse
